shifted includes to avoid cyclic inclusion error
'#include "Renderer.h"' needed to be removed from header files for ib, vb, vao and shader and added to their bodies, with the include for their own respective header removed. This is because the renderer was including them and creating a loop. The inclusion of renderer was only for error checking and some opengl data, but safe to keep this in the renderer as we can access all the base graphics objects through renderer.h anyway.